added optional argument to training map cunstructor
added optional argument to training map cunstructor.

This is to add the option to build evaluator that only considers the results of the predicates in the truth partition, otherwise the default truth value is used to evaluate unmatched targets.

When createOnGet is set to false then the inferred predicates without truth values are latent.